fix(build): pass Double values to Math.ceil

+++ app/src/main/java/com/nzf/markdown/share/RenderedMarkdownPdfShare.kt
@@ -71,7 +71,9 @@ object RenderedMarkdownPdfShare {
             val scaledDocumentHeight = sourceHeight.toFloat() * scale
             val pageCount = Math.max(
                 1,
-                Math.ceil(scaledDocumentHeight / contentHeight.toFloat()).toInt()
+                Math.ceil(
+                    scaledDocumentHeight.toDouble() / contentHeight.toDouble()
+                ).toInt()
             )
 
             val pdfDocument = PdfDocument()
